[feature](nereids) Support analyze for test SSB (#10415)
Follow-up #10241, this PR go through parse and analyze the SSB and add this functions:

1. support parse parenthesizedExpression
2. support analyze LogicalAggregate and LogicalSort
3. replace the functionCall to UnboundFunction and BoundFunction
4. support sum aggregate funciton
5. fix the dead loop in the ExpressionRewriter
6. refine some code

+++ fe/fe-core/src/main/java/org/apache/doris/nereids/types/DoubleType.java
@@ -23,7 +23,7 @@ import org.apache.doris.catalog.Type;
  * Double data type in Nereids.
  */
 public class DoubleType extends FractionalType {
-    public static IntegerType INSTANCE = new IntegerType();
+    public static DoubleType INSTANCE = new DoubleType();
 
     @Override
     public Type toCatalogDataType() {
